More 64-bit fixes (MP) based on commit from ioquake/jedi-academy (jonathangray)

+++ codemp/game/g_local.h
@@ -1523,7 +1523,7 @@ int BotAIStartFrame( int time );
 extern	level_locals_t	level;
 extern	gentity_t		g_entities[MAX_GENTITIES];
 
-#define	FOFS(x) ((int)&(((gentity_t *)0)->x))
+#define	FOFS(x) ((size_t)&(((gentity_t *)0)->x))
 
 #define XCVAR_PROTO
 	#include "g_xcvar.h"
